requirements.txt
APScheduler==3.4.0
dj-database-url==0.4.2
Django==1.11.5
django-grappelli==2.10.1
django-tables2==1.16.0
gunicorn==19.7.1
psycopg2==2.7.3.1
pytz==2017.2
six==1.11.0
tzlocal==1.5.1
whitenoise==3.3.1
django-bootstrap4==0.0.4
django-filter==1.1.0
tablib==0.12.1
